fix avatar spawn bug
