+++ packages/engine/src/avatar/AvatarSpawnSystem.ts
@@ -1,5 +1,5 @@
 import { Entity } from '../ecs/classes/Entity'
-import { defineQuery, getComponent, hasComponent } from '../ecs/functions/ComponentFunctions'
+import { addComponent, defineQuery, getComponent, hasComponent } from '../ecs/functions/ComponentFunctions'
 import { TransformComponent } from '../transform/components/TransformComponent'
 import { SpawnPointComponent } from '../scene/components/SpawnPointComponent'
 import { Quaternion, Vector3 } from 'three'
@@ -9,6 +9,14 @@ import { World } from '../ecs/classes/World'
 import { NetworkWorldAction } from '../networking/functions/NetworkWorldAction'
 import matches from 'ts-matches'
 import { isClient } from '../common/functions/isClient'
+import { Engine } from '../ecs/classes/Engine'
+import { AudioTagComponent } from '../audio/components/AudioTagComponent'
+import { ShadowComponent } from '../scene/components/ShadowComponent'
+import { LocalInputTagComponent } from '../input/components/LocalInputTagComponent'
+import { FollowCameraComponent, FollowCameraDefaultValues } from '../camera/components/FollowCameraComponent'
+import { PersistTagComponent } from '../scene/components/PersistTagComponent'
+import { NetworkObjectComponent } from '../networking/components/NetworkObjectComponent'
+import { AvatarComponent } from './components/AvatarComponent'
 
 const randomPositionCentered = (area: Vector3) => {
   return new Vector3((Math.random() - 0.5) * area.x, (Math.random() - 0.5) * area.y, (Math.random() - 0.5) * area.z)
@@ -46,7 +54,29 @@ export class SpawnPoints {
 
 export default async function AvatarSpawnSystem(world: World): Promise<System> {
   world.receptors.push((action) => {
-    matches(action).when(NetworkWorldAction.spawnAvatar.matches, createAvatar)
+    matches(action).when(NetworkWorldAction.spawnAvatar.matches, (spawnAction) => {
+      if (isClient) {
+        /**
+         * When changing location via a portal, the local client entity will be
+         * defined when the new world dispatches this action, so ignore it
+         */
+        if (Engine.userId === spawnAction.userId && hasComponent(world.localClientEntity, AvatarComponent)) {
+          return
+        }
+      }
+
+      const entity = createAvatar(spawnAction)
+      if (isClient) {
+        addComponent(entity, AudioTagComponent, {})
+        addComponent(entity, ShadowComponent, { receiveShadow: true, castShadow: true })
+
+        if (spawnAction.userId === Engine.userId) {
+          addComponent(entity, LocalInputTagComponent, {})
+          addComponent(entity, FollowCameraComponent, FollowCameraDefaultValues)
+          addComponent(entity, PersistTagComponent, {})
+        }
+      }
+    })
   })
 
   if (isClient) {
